Separate serialization manager methods, use compiled lambdas to call manager read
4 us > 200 ns
